fix(security): Win32 API call tweaks across modules
One-off Win32 API compatibility fixes: EXECUTION_STATE has no is_none (check .0 == 0); CreateDesktopW takes *mut c_void for pDevMode/lpsa (use null_mut); SHAddToRecentDocs expects u32 flag and Option<*const c_void> data (cast SHARD_PATHA.0, wrap in Some); OsString::from_wide needs OsStringExt in scope; DBTF_MEDIA is u32 so cast to u16 for the dbcv_flags match; drop redundant inner unsafe blocks and unused mut / unused c_void / unused Duration imports.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ src-tauri/src/security/usb_watch.rs
@@ -210,7 +210,7 @@ fn windows_parse_usb_event(wparam: usize, lparam: usize) -> Option<UsbDeviceInfo
     }
 
     let volume = unsafe { &*(lparam as *const win::DEV_BROADCAST_VOLUME) };
-    if volume.dbcv_flags & DBTF_MEDIA == 0 {
+    if volume.dbcv_flags & (DBTF_MEDIA as u16) == 0 {
         // Not a media change (e.g., drive letter added).
     }
